Move the strip-content-length helper into test/utils
