+++ test/utils/strip-content-length.js
